# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Interview Guide Platform API",
    "description": "High-performance Rust Axum API for Interview Guide & Technical Knowledge Platform",
    "contact": {
      "name": "Platform Engineering",
      "email": "eng@interviewguide.internal"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/decks/share": {
      "post": {
        "tags": [
          "Enterprise"
        ],
        "summary": "POST /api/decks/share",
        "description": "Creates a shareable custom deck with public share code (Item #14)",
        "operationId": "share_deck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareDeckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Custom deck published and share link generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDeckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/decks/shared/{share_code}": {
      "get": {
        "tags": [
          "Enterprise"
        ],
        "summary": "GET /api/decks/shared/:share_code",
        "operationId": "get_shared_deck",
        "parameters": [
          {
            "name": "share_code",
            "in": "path",
            "description": "8-character hex share code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shared deck details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDeckResponse"
                }
              }
            }
          },
          "404": {
            "description": "Shared deck not found"
          }
        }
      }
    },
...
    "/api/notes": {
      "get": {
        "tags": [
          "Notes"
        ],
        "operationId": "get_notes",
        "responses": {
          "200": {
            "description": "List all user notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NoteItem"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "operationId": "save_note",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Note updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteItem"
                }
              }
            }
          },
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/study/adaptive-next": {
      "get": {
        "tags": [
          "Progress"
        ],
        "summary": "GET /api/study/adaptive-next",
        "description": "Selects next question using Item Response Theory (IRT) (Item #37)",
        "operationId": "get_adaptive_next_question",
        "responses": {
          "200": {
            "description": "Next recommended question via IRT 2PL model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IrtAdaptiveQuestionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Category questions not found"
          }
        }
      }
    },
    "/api/study/sm18/review": {
      "post": {
        "tags": [
          "Spaced Repetition"
        ],
        "summary": "POST /api/study/sm18/review",
        "description": "Records spaced repetition review using SM-18 algorithm (Item #11)",
        "operationId": "persist_sm18_review",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sm18ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SM-18 spaced repetition review recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sm18ReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid rating"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IrtAdaptiveQuestionResponse": {
        "type": "object",
        "required": [
          "category_id",
          "question_number",
          "user_theta_ability",
          "question_difficulty_beta",
          "discrimination_alpha",
          "title",
          "target_competency"
        ],
        "properties": {
          "category_id": {
            "type": "string"
          },
          "discrimination_alpha": {
            "type": "number",
            "format": "float"
          },
          "question_difficulty_beta": {
            "type": "number",
            "format": "float"
          },
          "question_number": {
            "type": "integer",
            "format": "int32"
          },
          "target_competency": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "user_theta_ability": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "NoteItem": {
        "type": "object",
        "required": [
          "id",
          "category_id",
          "question_number",
          "note_content",
          "updated_at"
        ],
        "properties": {
          "category_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "note_content": {
            "type": "string"
          },
          "question_number": {
            "type": "integer",
            "format": "int32"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NoteRequest": {
        "type": "object",
        "required": [
          "category_id",
          "question_number",
          "note_content"
        ],
        "properties": {
          "category_id": {
            "type": "string"
          },
          "note_content": {
            "type": "string"
          },
          "question_number": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ShareDeckRequest": {
        "type": "object",
        "required": [
          "title",
          "question_ids",
          "is_public"
        ],
        "properties": {
          "category_id": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "is_public": {
            "type": "boolean"
          },
          "question_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ShareDeckResponse": {
        "type": "object",
        "required": [
          "deck_id",
          "title",
          "share_code",
          "question_count",
          "share_url",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deck_id": {
            "type": "string"
          },
          "question_count": {
            "type": "integer",
            "minimum": 0
          },
          "share_code": {
            "type": "string"
          },
          "share_url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "Sm18ReviewRequest": {
        "type": "object",
        "required": [
          "category_id",
          "question_number",
          "grade"
        ],
        "properties": {
          "category_id": {
            "type": "string"
          },
          "difficulty_factor": {
            "type": "number",
            "format": "float",
            "description": "Estimated difficulty from candidate (1 to 10)",
            "nullable": true
          },
          "grade": {
            "type": "integer",
            "format": "int32",
            "description": "Recall grade: 0 (blackout) to 5 (flawless)",
            "minimum": 0
          },
          "question_number": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Sm18ReviewResponse": {
        "type": "object",
        "required": [
          "category_id",
          "question_number",
          "repetition",
          "interval_days",
          "stability",
          "retrievability",
          "difficulty",
          "next_review_at"
        ],
        "properties": {
          "category_id": {
            "type": "string"
          },
          "difficulty": {
            "type": "number",
            "format": "float"
          },
          "interval_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_review_at": {
            "type": "string",
            "format": "date-time"
          },
          "question_number": {
            "type": "integer",
            "format": "int32"
          },
          "repetition": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "retrievability": {
            "type": "number",
            "format": "float"
          },
          "stability": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
    }
  },
...
}